requirements.txt
asgiref==3.6.0
autopep8==2.0.1
black==23.1.0
bleach==6.0.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
clickhouse-connect==0.5.24
colorama==0.4.6
Deprecated==1.2.14
distro==1.8.0
dj-database-url==1.2.0
Django==4.1.6
django-cors-headers==3.13.0
django-cron==0.6.0
django-crontab==0.7.1
django-db-logger==0.1.12
djangorestframework==3.14.0
drf-api-logger==1.1.12
forex-python==1.8
gevent==22.10.2
greenlet==2.0.2
gunicorn==20.1.0
idna==3.4
lz4==4.3.2
mypy-extensions==1.0.0
netaddr==0.8.0
numpy==1.24.3
packaging==23.0
pandas==2.0.1
pathspec==0.11.0
platformdirs==2.6.2
prometheus-client==0.16.0
#psycopg2==2.9.5
psycopg2-binary==2.9.5
pycodestyle==2.10.0
pycparser==2.21
python-dateutil==2.8.2
python-dotenv==0.21.1
pytz==2022.7.1
requests==2.28.2
simplejson==3.19.1
six==1.16.0
sqlparse==0.4.3
tzdata==2022.7
urllib3==1.26.14
webencodings==0.5.1
wrapt==1.15.0
yookassa==2.3.6
zope.event==4.6
zope.interface==6.0
zstandard==0.21.0
